+++ android/src/main/kotlin/es/antonborri/home_widget/HomeWidgetBackgroundReceiver.kt
@@ -1,0 +1,15 @@
+package es.antonborri.home_widget
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+import io.flutter.FlutterInjector
+
+class HomeWidgetBackgroundReceiver : BroadcastReceiver() {
+    override fun onReceive(context: Context, intent: Intent) {
+        val flutterLoader = FlutterInjector.instance().flutterLoader()
+        flutterLoader.startInitialization(context)
+        flutterLoader.ensureInitializationComplete(context, null)
+        HomeWidgetBackgroundService.enqueueWork(context, intent)
+    }
+}