+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 import 'dart:io';
+import 'dart:math';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -7,27 +8,56 @@ import 'package:flutter/services.dart';
 import 'package:home_widget/home_widget.dart';
 import 'package:workmanager/workmanager.dart';
 
+/// Used for Background Updates using Workmanager Plugin
 void callbackDispatcher() {
   Workmanager.executeTask((taskName, inputData) {
-    debugPrint('Native called background task: $taskName');
-
     final now = DateTime.now();
     return Future.wait<bool>([
-      HomeWidget.saveWidgetData('title', 'Updated from Background'),
-      HomeWidget.saveWidgetData('message',
-          '${now.year}-${now.month.toString().padLeft(2, '0')}-${now.day.toString().padLeft(2, '0')} ${now.hour.toString().padLeft(2, '0')}:${now.minute.toString().padLeft(2, '0')}:${now.second.toString().padLeft(2, '0')}'),
+      HomeWidget.saveWidgetData(
+        'title',
+        'Updated from Background',
+      ),
+      HomeWidget.saveWidgetData(
+        'message',
+        '${now.hour.toString().padLeft(2, '0')}:${now.minute.toString().padLeft(2, '0')}',
+      ),
       HomeWidget.updateWidget(
-          name: 'HomeWidgetExampleProvider', iOSName: 'HomeWidgetExample'),
+        name: 'HomeWidgetExampleProvider',
+        iOSName: 'HomeWidgetExample',
+      ),
     ]).then((value) {
       return !value.contains(false);
     });
   });
 }
 
+/// Called when Doing Background Work initiated from Widget
+void backgroundCallback(Uri data) async {
+  print(data);
+
+  if (data.host == 'titleclicked') {
+    final greetings = [
+      'Hello',
+      'Hallo',
+      'Bonjour',
+      'Hola',
+      'Ciao',
+      '哈洛',
+      '안녕하세요',
+      'xin chào'
+    ];
+    final selectedGreeting = greetings[Random().nextInt(greetings.length)];
+
+    await HomeWidget.saveWidgetData<String>('title', selectedGreeting);
+    await HomeWidget.updateWidget(
+        name: 'HomeWidgetExampleProvider', iOSName: 'HomeWidgetExample');
+  }
+}
+
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
   Workmanager.initialize(callbackDispatcher, isInDebugMode: kDebugMode);
-  runApp(MyApp());
+  runApp(MaterialApp(home: MyApp()));
 }
 
 class MyApp extends StatefulWidget {
@@ -43,6 +73,14 @@ class _MyAppState extends State<MyApp> {
   void initState() {
     super.initState();
     HomeWidget.setAppGroupId('YOUR_GROUP_ID');
+    HomeWidget.registerBackgroundCallback(backgroundCallback);
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    _checkForWidgetLaunch();
+    HomeWidget.widgetClicked.listen(_launchedFromWidget);
   }
 
   @override
@@ -91,6 +129,21 @@ class _MyAppState extends State<MyApp> {
     await _updateWidget();
   }
 
+  void _checkForWidgetLaunch() {
+    HomeWidget.initiallyLaunchedFromHomeWidget().then(_launchedFromWidget);
+  }
+
+  void _launchedFromWidget(Uri uri) {
+    if (uri != null) {
+      showDialog(
+          context: context,
+          builder: (buildContext) => AlertDialog(
+                title: Text('App started from HomeScreenWidget'),
+                content: Text('Here is the URI: $uri'),
+              ));
+    }
+  }
+
   void _startBackgroundUpdate() {
     Workmanager.registerPeriodicTask('1', 'widgetBackgroundUpdate',
         frequency: Duration(minutes: 15));
@@ -102,46 +155,48 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('HomeWidget Example'),
-        ),
-        body: Center(
-          child: Column(
-            children: [
-              TextField(
-                decoration: InputDecoration(
-                  hintText: 'Title',
-                ),
-                controller: _titleController,
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('HomeWidget Example'),
+      ),
+      body: Center(
+        child: Column(
+          children: [
+            TextField(
+              decoration: InputDecoration(
+                hintText: 'Title',
               ),
-              TextField(
-                decoration: InputDecoration(
-                  hintText: 'Body',
-                ),
-                controller: _messageController,
+              controller: _titleController,
+            ),
+            TextField(
+              decoration: InputDecoration(
+                hintText: 'Body',
               ),
+              controller: _messageController,
+            ),
+            ElevatedButton(
+              onPressed: _sendAndUpdate,
+              child: Text('Send Data to Widget'),
+            ),
+            ElevatedButton(
+              onPressed: _loadData,
+              child: Text('Load Data'),
+            ),
+            ElevatedButton(
+              onPressed: _checkForWidgetLaunch,
+              child: Text('Check For Widget Launch'),
+            ),
+            if (Platform.isAndroid)
               ElevatedButton(
-                onPressed: _sendAndUpdate,
-                child: Text('Send Data to Widget'),
+                onPressed: _startBackgroundUpdate,
+                child: Text('Update in background'),
               ),
+            if (Platform.isAndroid)
               ElevatedButton(
-                onPressed: _loadData,
-                child: Text('Load Data'),
-              ),
-              if (Platform.isAndroid)
-                ElevatedButton(
-                  onPressed: _startBackgroundUpdate,
-                  child: Text('Update in background'),
-                ),
-              if (Platform.isAndroid)
-                ElevatedButton(
-                  onPressed: _stopBackgroundUpdate,
-                  child: Text('Stop updating in background'),
-                )
-            ],
-          ),
+                onPressed: _stopBackgroundUpdate,
+                child: Text('Stop updating in background'),
+              )
+          ],
         ),
       ),
     );