+++ ios/Classes/SwiftHomeWidgetPlugin.swift
@@ -2,11 +2,34 @@ import Flutter
 import UIKit
 import WidgetKit
 
-public class SwiftHomeWidgetPlugin: NSObject, FlutterPlugin {
+public class SwiftHomeWidgetPlugin: NSObject, FlutterPlugin, FlutterStreamHandler {
+    
+    private static var groupId : String?
+    
+    private var initialUrl : URL?
+    private var latestUrl : URL? {
+        didSet {
+            if(latestUrl != nil) {
+                eventSink?.self(latestUrl?.absoluteString)
+            }
+        }
+    }
+    
+    private var eventSink : FlutterEventSink?
+    
+    private let notInitializedError = FlutterError(
+    code: "-7", message: "AppGroupId not set. Call setAppGroupId first", details: nil)
+    
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let channel = FlutterMethodChannel(name: "home_widget", binaryMessenger: registrar.messenger())
         let instance = SwiftHomeWidgetPlugin()
+        
+        let channel = FlutterMethodChannel(name: "home_widget", binaryMessenger: registrar.messenger())
         registrar.addMethodCallDelegate(instance, channel: channel)
+        
+        let eventChannel = FlutterEventChannel(name: "home_widget/updates", binaryMessenger: registrar.messenger())
+        eventChannel.setStreamHandler(instance)
+        
+        registrar.addApplicationDelegate(instance)
     }
     
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
@@ -55,7 +78,7 @@ public class SwiftHomeWidgetPlugin: NSObject, FlutterPlugin {
                 result(FlutterError(code: "-2", message: "InvalidArguments getWidgetData must be called with id", details: nil))
             }
         } else if call.method == "updateWidget" {
-            
+
             guard let args = call.arguments else {
                 return
             }
@@ -71,14 +94,49 @@ public class SwiftHomeWidgetPlugin: NSObject, FlutterPlugin {
             } else {
                 result(FlutterError(code: "-3", message: "InvalidArguments updateWidget must be called with name", details: nil))
             }
+        } else if call.method == "initiallyLaunchedFromHomeWidget" {
+            if (SwiftHomeWidgetPlugin.groupId == nil) {
+                result(notInitializedError)
+                return
+            }
+            if(initialUrl != nil) {
+                result(initialUrl?.absoluteString)
+            }
         } else {
             result(FlutterMethodNotImplemented)
         }
     }
     
-    private let notInitializedError = FlutterError(
-    code: "-7", message: "AppGroupId not set. Call setAppGroupId first", details: nil)
+    public func onListen(withArguments arguments: Any?, eventSink events: @escaping FlutterEventSink) -> FlutterError? {
+        eventSink = events
+        return nil
+    }
     
+    public func onCancel(withArguments arguments: Any?) -> FlutterError? {
+        eventSink = nil
+        return nil
+    }
+
     
-    private static var groupId : String?
+    public func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [AnyHashable : Any] = [:]) -> Bool {
+        let launchUrl = (launchOptions[UIApplication.LaunchOptionsKey.url] as? NSURL)?.absoluteURL
+        if(launchUrl != nil && isWidgetUrl(url: launchUrl!)) {
+            initialUrl = launchUrl?.absoluteURL
+            latestUrl = initialUrl
+        }
+        return true
+    }
+    
+    public func application(_ application: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
+        if(isWidgetUrl(url: url)) {
+            latestUrl = url
+        }
+        return true
+    }
+    
+    private func isWidgetUrl(url: URL) -> Bool {
+        let components = URLComponents.init(url: url, resolvingAgainstBaseURL: false)
+        return components?.queryItems?.contains(where: {(item) in item.name == "homeWidget"}) ?? false
+    
+    }
 }